+++ src/main/java/com/example/backend/oauth2/dto/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.example.backend.Oauth2.utils;
+package com.example.backend.oauth2.dto;
 
 
 import lombok.Builder;
@@ -6,13 +6,13 @@ import lombok.Getter;
 
 @Builder
 @Getter
-public class KakaoUserInfo {
+public class GoogleUserInfo {
 
     private String email;
     private String nickname;
     private String provider;
 
-    public KakaoUserInfo(String email, String nickname, String provider) {
+    public GoogleUserInfo(String email, String nickname, String provider) {
         this.email = email;
         this.nickname = nickname;
         this.provider = provider;
