+++ src/main/java/com/example/backend/oauth2/dto/KakaoUserInfo.java
@@ -1,18 +1,18 @@
-package com.example.backend.Oauth2.utils;
+package com.example.backend.oauth2.dto;
+
 
 import lombok.Builder;
 import lombok.Getter;
-import org.springframework.http.ResponseEntity;
-
 
 @Builder
 @Getter
-public class NaverUserInfo{
+public class KakaoUserInfo {
+
     private String email;
     private String nickname;
     private String provider;
 
-    public NaverUserInfo(String email, String nickname, String provider) {
+    public KakaoUserInfo(String email, String nickname, String provider) {
         this.email = email;
         this.nickname = nickname;
         this.provider = provider;
