+++ src/main/java/com/example/backend/oauth2/dto/NaverUserInfo.java
@@ -1,18 +1,17 @@
-package com.example.backend.Oauth2.utils;
-
+package com.example.backend.oauth2.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
+
 @Builder
 @Getter
-public class GoogleUserInfo {
-
+public class NaverUserInfo{
     private String email;
     private String nickname;
     private String provider;
 
-    public GoogleUserInfo(String email, String nickname, String provider) {
+    public NaverUserInfo(String email, String nickname, String provider) {
         this.email = email;
         this.nickname = nickname;
         this.provider = provider;
