+++ src/main/java/com/example/backend/oauth2/service/NaverOauth2Service.java
@@ -1,7 +1,6 @@
-package com.example.backend.Oauth2.service;
+package com.example.backend.oauth2.service;
 
-import com.example.backend.Oauth2.utils.GoogleUserInfo;
-import com.example.backend.Oauth2.utils.KakaoUserInfo;
+import com.example.backend.oauth2.dto.NaverUserInfo;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -13,41 +12,32 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
-import java.util.HashMap;
-import java.util.Map;
-
 
 @Service
 @PropertySource("classpath:application.properties")
-public class GoogleOauth2Service implements CustomOauth2Service {
+public class NaverOauth2Service implements CustomOauth2Service{
 
+    @Value("${naver.token-url}")
+    private String NAVER_TOKEN_URL;
+    @Value("${naver.client-id}")
+    private String NAVER_CLIENT_ID;
+    @Value("${naver.client-secret}")
+    private String NAVER_CLIENT_SECRET;
 
-    @Value("${google.token-url}")
-    private String GOOGLE_TOKEN_URL;
-    @Value("${google.client-id}")
-    private String GOOGLE_CLIENT_ID;
-    @Value("${google.client-secret}")
-    private String GOOGLE_CLIENT_SECRET;
-    @Value("${google.redirect-url}")
-    private String GOOGLE_REDIRECT_URL;
+    public NaverUserInfo getNaverUserInfo(ResponseEntity<String> response) throws JsonProcessingException {
 
-
-
-
-    public GoogleUserInfo getGoogleUserInfo(ResponseEntity<String> response) throws JsonProcessingException {
         String responseBody = response.getBody();
         ObjectMapper objectMapper = new ObjectMapper();
         JsonNode jsonNode = objectMapper.readTree(responseBody);
-        String nickname = jsonNode.get("name").asText();
-        String email = jsonNode.get("email").asText();
+        String nickname = jsonNode.get("response").get("name").asText();
+        String email = jsonNode.get("response").get("email").asText();
 
-        return new GoogleUserInfo(email,nickname, "google");
+        return new NaverUserInfo(email,nickname, "naver");
     }
 
-    //google code로 accessToken 받아오기
+
     @Override
     public ResponseEntity<String> getUserProfile(String code) {
-
         RestTemplate restTemplate = new RestTemplate();
         String accessCode = getAccessToken(code);
         HttpHeaders headers = new HttpHeaders();
@@ -58,8 +48,8 @@ public class GoogleOauth2Service implements CustomOauth2Service {
         HttpEntity<MultiValueMap<String,String>> request = new HttpEntity<>(headers);
 
         ResponseEntity<String> response = restTemplate.exchange(
-                "https://www.googleapis.com/oauth2/v2/userinfo",
-                HttpMethod.GET,
+                "https://openapi.naver.com/v1/nid/me",
+                HttpMethod.POST,
                 request,
                 String.class
         );
@@ -68,21 +58,30 @@ public class GoogleOauth2Service implements CustomOauth2Service {
             return response;
         }
         return null;
-
     }
 
-    @Override
+
+
     public String getAccessToken(String code) {
         RestTemplate restTemplate = new RestTemplate();
-        Map<String, String> params = new HashMap<>();
 
-        params.put("code", code);
-        params.put("client_id", GOOGLE_CLIENT_ID);
-        params.put("client_secret", GOOGLE_CLIENT_SECRET);
-        params.put("redirect_uri", GOOGLE_REDIRECT_URL);
-        params.put("grant_type", "authorization_code");
+        // 헤더 설정
+        //kakao는 application/x-www-form-urlencoded
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
 
-        ResponseEntity<String> response = restTemplate.postForEntity(GOOGLE_TOKEN_URL, params, String.class);
+        // 바디 설정
+        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+        params.add("grant_type", "authorization_code");
+        params.add("client_id", NAVER_CLIENT_ID);
+        params.add("client_secret", NAVER_CLIENT_SECRET);
+        params.add("code", code);
+
+        // 요청 엔티티 생성
+        HttpEntity<MultiValueMap<String, String>> requestEntity = new HttpEntity<>(params, headers);
+
+        // 요청 전송
+        ResponseEntity<String> response = restTemplate.exchange(NAVER_TOKEN_URL, HttpMethod.POST, requestEntity, String.class);
 
         if(response.getStatusCode() == HttpStatus.OK){
             try {
