+++ src/main/resources/mapper/UserMapper.xml
@@ -11,7 +11,7 @@
     </insert>
 
     <!-- 기존 사용자 로그인 : 주어진 이메일 사용해 기존 사용자 존재 여부 확인 -->
-    <select id="findUserByEmail" parameterType="long" resultType="com.example.backend.Users.vo.UserVO">
+    <select id="findUserByEmail" parameterType="string" resultType="com.example.backend.Users.vo.UserVO">
         SELECT * FROM Users WHERE email = #{email}
     </select>
